+++ echelon/active.php
@@ -67,7 +67,10 @@ require 'app/views/global/header.php';
 
 if(!$db->error) :
 ?>
-
+<div class="page-header">
+	<h1>Inactive Admin</h1>
+	<p><span class="badge"><?= $total_rows ?></span> admins who have not been seen by B3 for</p>
+</div>
 <table class="table table-striped table-hover">
 	<form class="form-inline" action="active.php" method="get">
 		<div class="form-group">
@@ -84,7 +87,6 @@ if(!$db->error) :
 			</select>
 		</div>
 	</form>
-	<caption>Inactive Admins<small>There are <strong><?php echo $total_rows; ?></strong> admins who have not been seen by B3 for</small></caption>
 	<thead>
 		<tr>
 			<th>Name